+++ pages/index.tsx
@@ -4,7 +4,7 @@ import Hero from "../components/hero";
 import Drops from "../components/drops";
 import Faq from "../components/faq";
 import WhatIsItAbout from "../components/whats-about";
-import Header from "./header";
+import Header from "../components/header";
 
 const Home: NextPage = () => {
   return (
@@ -14,12 +14,17 @@ const Home: NextPage = () => {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link href="https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&display=swap" rel="stylesheet" />
+        <link
+          href="https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&display=swap"
+          rel="stylesheet"
+        />
       </Head>
-      <Header/>
-      <Hero/>
+      <div className="headerContainer absolute w-[100%]">
+        <Header />
+      </div>
+      <Hero />
       <WhatIsItAbout />
-      <Drops/>
+      <Drops />
       <Faq />
     </div>
   );
